retry, correction: Makefile, EOF

+++ cpp_00/ex00/megaphone.cpp
@@ -23,7 +23,7 @@ int main(int argc, char **argv){
 	}
 	for (int i = 1; i <argc; i++){
 		str = argv[i];
-		for (int j = 0; j < str.length(); j++)
+		for (size_t j = 0; j < str.length(); j++)
 			std::cout << (char)toupper(str[j]);
 	}
 	std::cout << std::endl;
